pyproject.toml
[project]
name = "strategy-lib"
version = "0.1.0"
description = "多市场量化策略研究库：因子研究、回测、策略开发"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "0xBroleez" }]
license = { text = "MIT" }

dependencies = [
    # 核心数值与数据
    "numpy>=1.26",
    "pandas>=2.1",
    "scipy>=1.11",
    "statsmodels>=0.14",
    "pyarrow>=15.0",          # parquet 缓存
    "pyyaml>=6.0",

    # 回测引擎
    "vectorbt>=0.26",

    # 技术指标
    "pandas-ta>=0.3.14b",

    # 数据源
    "ccxt>=4.2",              # Crypto
    "yfinance>=0.2.40",       # 美股 / ETF
    "akshare>=1.13",          # A股 / 港股 / ETF（免费、无 token）

    # 可视化
    "matplotlib>=3.8",
    "plotly>=5.20",
    "seaborn>=0.13",

    # 工具
    "tqdm>=4.66",
    "click>=8.1",             # CLI
    "loguru>=0.7",            # 日志
]

[project.optional-dependencies]
notebook = [
    "jupyterlab>=4.0",
    "ipykernel>=6.29",
    "ipywidgets>=8.1",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
slib = "strategy_lib.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/strategy_lib"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]   # 长行交给 formatter

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
